Split code to modules
Moves are mostly done w/o changes. There are APIs (parameters and return
values) that are subject to further refactoring.

Removed duplicate code in resolve the function for multihead
connections.

Part of #166.
