+++ graphql/convert_schema/init.lua
@@ -1,0 +1,7 @@
+local schema = require('graphql.convert_schema.schema')
+
+local convert_schema = {}
+
+convert_schema.convert = schema.convert
+
+return convert_schema
